Update software-submit-form.php
The software slug is now obtained from parsing the zip file from the download link. (That file is temporarily uploaded and then deleted.) The current version of the software is now obtained from parsing the download link URL.

The fields in the form which previously required manual entry of these pieces of information are now redundant and so have been deleted, along with the server-side checks that those fields had been filled.